refactor(wxThread): remove private thread class from headers
